--- a/linear-base.cabal
+++ b/linear-base.cabal
@@ -219,6 +219,7 @@ benchmark bench
     other-modules:
         Data.Mutable.HashMap
         Data.Mutable.Array
+        Data.Mutable.Quicksort
     default-language: Haskell2010
     build-depends:
         base,