So far:

```
9.4.8 resolver: lts-21.25:

  quicksort
    size 1000
      linArrayQuicksort:                                                         OK
        773  μs ±  65 μs
      lazyListQuicksort:                                                         OK
        276  μs ±  16 μs
      stdLibSort:                                                                OK
        219  μs ±  12 μs
    size 50000
      linArrayQuicksort:                                                         OK
        63.9 ms ± 3.2 ms
      lazyListQuicksort:                                                         OK
        32.4 ms ± 2.3 ms
      stdLibSort:                                                                OK
        38.1 ms ± 1.9 ms
    size 1000000
      linArrayQuicksort:                                                         OK
        3.557 s ± 322 ms
      lazyListQuicksort:                                                         OK
        2.135 s ±  72 ms
      stdLibSort:                                                                OK
        4.309 s ± 212 ms

9.8.2 resolver: nightly-2024-09-10:

  quicksort              
    size 1000            
      linArrayQuicksort:                                                         OK
        221  μs ± 5.6 μs 
      lazyListQuicksort:                                                         OK
        455  μs ±  38 μs 
      stdLibSort:                                                                OK
        362  μs ±  11 μs 
    size 50000           
      linArrayQuicksort:                                                         OK
        22.1 ms ± 495 μs 
      lazyListQuicksort:                                                         OK
        82.8 ms ± 4.6 ms 
      stdLibSort:                                                                OK
        145  ms ±  11 ms 
    size 1000000         
      linArrayQuicksort:                                                         OK
        718  ms ±  64 ms 
      lazyListQuicksort:                                                         OK
        2.138 s ±  47 ms 
      stdLibSort:                                                                OK
        4.241 s ±  96 ms
```

I didn't have time to benchmark 9.10 yet (because it must use `cabal --allow-newer`; there is no stackage set available), but there definitely seems to be an improvement in 9.8 compared to 9.4 (especially for the linear one, but also for the other variants albeit to a lesser extent).