@tbagrel1 2024-09-11
@aspiwack After some struggle with the CI, and some ugly fix [here](https://github.com/tweag/linear-base/pull/477/files#diff-b8ed06757045fac949c89f2139a862498ad2b6d1f82c61a31e7c91d6cf0eaa70R3), I think the PR is finally ready for review.