@tbagrel1 2024-09-13
With this PR, I enable benchmarks in the cabal matrix of the CI, so that we can check performance across versions of linear-base and versions of GHC easily.

The thing is, `stack` is our preferred platform, and we use stackage for package version choice and build plan. We partially support `cabal` on the project, but we have little constraint bounds/overrides to help cabal find a proper build plan.

Without any options, `cabal build linear-base:bench:bench` doesn't succeed with GHC 9.8/GHC 9.10 as of now. One thing we used in the past is to use `--allow-newer`, and it is enough to help the constraint solver to find a build plan for the library, but not for the benchmarks, mainly because of a constraint on `filepath` in transitive dependencies.

In https://github.com/tweag/linear-base/pull/477/commits/8d1533e4754e9b224026f685ce83e65a12b30ba0, I pinned the version of `filepath` and the package state index. It is enough for the constraint solver to succeed, and shouldn't break unexpectedly as long as this version of filepath is compatible with other constraints. 

By some very (un)fortunate coincidence, @amesgen & I found that if we enable tests (`--enable-tests` or `tests: True` in `cabal.project`) at the moment of building the benchmark, it helps the constraint solver to succeed, with no need to pin `filepath`. It might be more brittle or less brittle than the other solution, I don't know. This option is presented in https://github.com/tweag/linear-base/pull/477/commits/919fc6c2fe1e44ae912efaabdc57090304e9d0f2