@tbagrel1 2024-09-18
For reference, here is the comparison of generated Core between GHC 9.6 and GHC 9.8 for quicksort. It seems that `Ur` are indeed unboxed in GHC 9.8 while they aren't in GHC 9.6:

**GHC 9.6**
```
partition_$s$wpartition2
  = \ sc sc1 ww ww1 ->
      case $w$c<2 ww1 ww of {
        False ->
          case sc1 of { Array ww2 ->
          case ($wassertIndexInRange (lvl1 `cast` <Co:4> :: ...) ww ww2)
               `cast` <Co:2> :: ...
          of nt
          { __DEFAULT ->
          case $wget ww (nt `cast` <Co:3> :: ...) of { (# ww3, ww4 #) ->
          case ww3 of { Ur lVal ->
          case ($wassertIndexInRange (lvl3 `cast` <Co:4> :: ...) ww1 ww4)
               `cast` <Co:2> :: ...
          of nt1
          { __DEFAULT ->
          case $wget ww1 (nt1 `cast` <Co:3> :: ...) of { (# ww5, ww6 #) ->
          case ww5 of { Ur rVal ->
          case lVal of { I# ww7 ->
[...]
```

**GHC 9.8**

```
partition_$s$wpartition2
  = \ sc sc1 ww ww1 ->
      case <# ww1 ww of {
        __DEFAULT ->
          case <=# 0# ww of {
            __DEFAULT -> case partition17 of wild { };
            1# ->
              case <# ww (sizeofMutableArray# (sc1 `cast` <Co:2> :: ...)) of {
                __DEFAULT -> case partition17 of wild { };
                1# ->
                  case $wget ww sc1 of { (# ww2, ww3 #) ->
                  case <=# 0# ww1 of {
                    __DEFAULT -> case partition16 of wild { };
                    1# ->
                      case <# ww1 (sizeofMutableArray# (ww3 `cast` <Co:2> :: ...)) of {
                        __DEFAULT -> case partition16 of wild { };
                        1# ->
                          case $wget ww1 ww3 of { (# ww5, ww6 #) ->
                          case ww2 of { I# f1 ->
                          join {
                            $j
                              = case ww5 of { I# f2 ->
[Quicksort-GHC-9.6-full.dump-simpl.txt](https://github.com/user-attachments/files/17040762/Quicksort-GHC-9.6-full.dump-simpl.txt)
[Quicksort-GHC-9.6-suppr.dump-simpl.txt](https://github.com/user-attachments/files/17040765/Quicksort-GHC-9.6-suppr.dump-simpl.txt)
[Quicksort-GHC-9.8-full.dump-simpl.txt](https://github.com/user-attachments/files/17040766/Quicksort-GHC-9.8-full.dump-simpl.txt)
[Quicksort-GHC-9.8-suppr.dump-simpl.txt](https://github.com/user-attachments/files/17040767/Quicksort-GHC-9.8-suppr.dump-simpl.txt)

```