@aspiwack 2024-09-19
I'll be frank: I'm not sure why unboxing of `Ur` starts appearing specifically in GHC 9.8. Maybe I did something. But I don't remember. (I remember there was a refactor of the worker/wrapper split code at some point which I believed to accidentally enable unboxing of `Ur`, but I seem to remember that it was in effect in 9.6 already. It doesn't really matter, as long as this is good now)