@tbagrel1 2024-09-20
I discovered that the `test`, `bench`, etc. subdirectories didn't match the `src` structure.

I added a commit to address this, but because the change would affect a dozen of files, I'm requesting your approval again :) Let me know if you think a separate PR would be better.